refactor: move app-initializers into their respective modules
