+++ src/app/app.module.ts
@@ -17,15 +17,8 @@
 
 import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import {
-  APP_INITIALIZER,
-  ErrorHandler,
-  Inject,
-  LOCALE_ID,
-  NgModule,
-} from "@angular/core";
+import { LOCALE_ID, NgModule } from "@angular/core";
 import { HttpClientModule } from "@angular/common/http";
-import * as Sentry from "@sentry/angular";
 
 import { AppComponent } from "./app.component";
 import { allRoutes } from "./app.routing";
@@ -40,8 +33,6 @@ import {
 import { environment } from "../environments/environment";
 import { AnalyticsService } from "./core/analytics/analytics.service";
 import { ConfigurableEnumModule } from "./core/basic-datatypes/configurable-enum/configurable-enum.module";
-import { MatPaginatorIntl } from "@angular/material/paginator";
-import { TranslatableMatPaginator } from "./core/language/TranslatableMatPaginator";
 import { FaIconLibrary } from "@fortawesome/angular-fontawesome";
 import { fas } from "@fortawesome/free-solid-svg-icons";
 import { far } from "@fortawesome/free-regular-svg-icons";
@@ -81,13 +72,11 @@ import { HistoricalDataModule } from "./features/historical-data/historical-data
 import { MatchingEntitiesModule } from "./features/matching-entities/matching-entities.module";
 import { ProgressDashboardWidgetModule } from "./features/dashboard-widgets/progress-dashboard-widget/progress-dashboard-widget.module";
 import { ReportingModule } from "./features/reporting/reporting.module";
-import { Router, RouterModule } from "@angular/router";
+import { RouterModule } from "@angular/router";
 import { TodosModule } from "./features/todos/todos.module";
-import moment from "moment";
-import { getLocaleFirstDayOfWeek } from "@angular/common";
 import { waitForChangeTo } from "./core/session/session-states/session-utils";
 import { LoginState } from "./core/session/session-states/login-state.enum";
-import { appInitializers } from "./app-initializers";
+import { APP_INITIALIZER_PROPAGATE_CONFIG_UPDATES } from "./core/config/config.app-initializer";
 import { ImportModule } from "./core/import/import.module";
 import { ShortcutDashboardWidgetModule } from "./features/dashboard-widgets/shortcut-dashboard-widget/shortcut-dashboard-widget.module";
 import { EntityCountDashboardWidgetModule } from "./features/dashboard-widgets/entity-count-dashboard-widget/entity-count-dashboard-widget.module";
@@ -95,6 +84,8 @@ import { BirthdayDashboardWidgetModule } from "./features/dashboard-widgets/birt
 import { MarkdownPageModule } from "./features/markdown-page/markdown-page.module";
 import { LoginStateSubject } from "./core/session/session-type";
 import { AdminModule } from "./core/admin/admin.module";
+import { Logging } from "./core/logging/logging.service";
+import { APP_INITIALIZER_DEMO_DATA } from "./core/demo-data/demo-data.app-initializer";
 
 /**
  * Main entry point of the application.
@@ -147,23 +138,7 @@ import { AdminModule } from "./core/admin/admin.module";
     MatDialogModule,
   ],
   providers: [
-    /* Sentry setup */
-    {
-      provide: ErrorHandler,
-      useValue: Sentry.createErrorHandler(),
-    },
-    {
-      provide: Sentry.TraceService,
-      deps: [Router],
-    },
-    {
-      provide: APP_INITIALIZER,
-      useFactory: () => () => {},
-      deps: [Sentry.TraceService],
-      multi: true,
-    },
-
-    { provide: MatPaginatorIntl, useValue: TranslatableMatPaginator() },
+    ...Logging.getAngularTracingProviders(),
     { provide: ComponentRegistry, useValue: componentRegistry },
     { provide: EntityRegistry, useValue: entityRegistry },
     { provide: WINDOW_TOKEN, useValue: window },
@@ -190,17 +165,13 @@ import { AdminModule } from "./core/admin/admin.module";
       }),
       deps: [LoginStateSubject],
     },
-    appInitializers,
+    APP_INITIALIZER_PROPAGATE_CONFIG_UPDATES,
+    APP_INITIALIZER_DEMO_DATA,
   ],
   bootstrap: [AppComponent],
 })
 export class AppModule {
-  constructor(icons: FaIconLibrary, @Inject(LOCALE_ID) locale: string) {
+  constructor(icons: FaIconLibrary) {
     icons.addIconPacks(fas, far);
-    moment.updateLocale(moment.locale(), {
-      week: {
-        dow: getLocaleFirstDayOfWeek(locale),
-      },
-    });
   }
 }
