+++ src/app/core/import/import-entity-type/import-entity-type.stories.ts
@@ -2,14 +2,17 @@ import { applicationConfig, Meta, StoryFn } from "@storybook/angular";
 import { StorybookBaseModule } from "../../../utils/storybook-base.module";
 import { ImportEntityTypeComponent } from "./import-entity-type.component";
 import { importProvidersFrom } from "@angular/core";
-import { appInitializers } from "../../../app-initializers";
+import { APP_INITIALIZER_PROPAGATE_CONFIG_UPDATES } from "../../config/config.app-initializer";
 
 export default {
   title: "Features/Import/2 Select Entity Type",
   component: ImportEntityTypeComponent,
   decorators: [
     applicationConfig({
-      providers: [importProvidersFrom(StorybookBaseModule), appInitializers],
+      providers: [
+        importProvidersFrom(StorybookBaseModule),
+        APP_INITIALIZER_PROPAGATE_CONFIG_UPDATES,
+      ],
     }),
   ],
 } as Meta;
